Add Membership tests
Expect to see failures (fixed in next commits)

Add tests for issue #27.
For a existing membership (both Current & Grace) where payment was not GC and renewal is GC,
the dates and status should not be changed by setting up the mandate.

Modify testWebhookPaymentConfirmationFirst() to be more specific about dates
Not sure whether this is the _desired_ behaviour re dates, but it documents
the _current_ behaviour which should help if we want to change.